fsutil: move many things out of the header into cpp
