+++ src/flit.h
@@ -91,6 +91,7 @@
 #include "MpiEnvironment.h"
 #include "TestBase.h"
 #include "flitHelpers.h"
+#include "fsutil.h"
 
 #include <algorithm>
 #include <chrono>
@@ -192,9 +193,6 @@ static std::string info =
   "  Optimization level: \"" FLIT_OPTL     "\"\n"
   "  Compiler flags:     \"" FLIT_SWITCHES "\"\n"
   "  Filename:           \"" FLIT_FILENAME "\"\n";
-
-/// Read file contents entirely into a string
-std::string readFile(const std::string &filename);
 
 /// Parse the results file into a vector of results
 std::vector<TestResult> parseResults(std::istream &in);
@@ -382,8 +380,8 @@ long double runComparison_impl(TestFactory* factory, const TestResult &gt,
       throw std::invalid_argument("baseline comparison type is not None when"
                                   " the resultfile is defined");
     }
-    Variant gtval = Variant::fromString(readFile(gt.resultfile()));
-    Variant resval = Variant::fromString(readFile(res.resultfile()));
+    Variant gtval = Variant::fromString(fsutil::readfile(gt.resultfile()));
+    Variant resval = Variant::fromString(fsutil::readfile(res.resultfile()));
     return test->variant_compare(gtval, resval);
   }
   return test->variant_compare(gt.result(), res.result());
